requirements.txt
asgiref==3.11.1
Django==6.0.2
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
python-dotenv==1.2.1
sqlparse==0.5.5
django-cors-headers==4.9.0
django-csp==4.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
djangorestframework==3.15.2
